requirements.txt
aiofiles==0.8.0
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cloudinary==1.28.0
cryptography==36.0.0
ecdsa==0.17.0
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
paginate==0.5.6
paginate-sqlalchemy==0.3.1
passlib==1.7.4
psycopg2==2.9.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.15.0
Werkzeug==2.0.2
